Refactor namespaces and cleanup unused dependencies
