+++ src/ShapeCrawler/ShapeCollection/GroupShapes/GroupedShapes.cs
@@ -3,14 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
-using ShapeCrawler.ShapeCollection.GroupShapes;
 using ShapeCrawler.Texts;
 using P = DocumentFormat.OpenXml.Presentation;
 using Picture = ShapeCrawler.Drawing.Picture;
 
-namespace ShapeCrawler.ShapeCollection;
+namespace ShapeCrawler.ShapeCollection.GroupShapes;
 
-internal sealed class GroupedShapes : IShapes
+internal sealed record GroupedShapes : IShapes
 {
     private readonly OpenXmlPart sdkTypedOpenXmlPart;
     private readonly IEnumerable<OpenXmlCompositeElement> pGroupElements;
