+++ src/ShapeCrawler/ShapeCollection/SlideShapes.cs
@@ -9,9 +9,9 @@ using System.Reflection;
 using System.Text.RegularExpressions;
 using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
+using ShapeCrawler.Charts;
 using ShapeCrawler.Exceptions;
 using ShapeCrawler.Extensions;
-using ShapeCrawler.Services;
 using ShapeCrawler.Shared;
 using ShapeCrawler.Units;
 using SkiaSharp;
